SNAP-187: Added many tests against the compiler and updated the compiler accordingly.
Added --scripts in the coverage script so it can run the compiler scripts.
Added code to print out decimal number with the number of digits after a period.
Moved the compile_at_keyword() handling flow to the replace_at_keyword() which happens at the same time as the variable replacement.
Fixed the lookup for an OPEN_CURLYBRACKET to only select those that are acceptable, i.e. {}-blocks of selectors.
Added code for the @include, did not test it yet.
Properly check the :lang() function parameter with the language and country names.
I moved the expression parser to its own files.
Fixed the replace_child() function so it properly inserts/deletes the right children.
Fixed the C++ comments ending (added the \n).
Added a few more node types.
Output strings with the correct ' and " escaped and quoted.
Displays the f_string of the PLACEHOLDER.
Fixed the nth-child parser, the spaces are to be returned otherwise "3 3" looks like "33".
Fixed the return of ODD to ODD instead of EVEN.
Added a new type named VARIABLE_FUNCTION.