SNAP-187: Continued work on compiler: mainly variables and more coverage.
Changed the state to directly search and set variables.
Prevent parsing the body of a variable function until it gets used (otherwise the variables inside that block are not going to be replaced by the proper values.)
Added a node::clone() because variables must be cloned and not just their pointer duplicated (with simple variables, no problems, but in many cases, not good!)
When saving a variable in a node (node::set_variable()) we save a LIST for the value with the variable information (i.e. the VARIABLE_FUNCTION) so we can properly call the function if their is one.
Properly parse the variable function parameters when calling the function.
Implemented the replacement of variable functions (and plain functions if applicable.)
Apply all expressions in a list of items in a declaration. This works as expected now, leaving whitespaces at the right place, etc.
Add support for font metrics in nodes. Still need to check for that type of entry in a declaration before to transform it with expression().
Fixed the ++idx while replacing variables, in case a change occur, we may be required to not do ++idx.
Removed the use of a cache for the validation scripts. We may be able to use the node::clone() there, but a direct use of the loaded parser data works only once.
Expressions now check a HASH node as a color, if not a color, an error is generated.