SNAP-187: Got the "current version" of the compiler 100% covered by the test.
Moved the scripts under a sub-directory so the @import have to reference those names making it possible to create libraries.
Added the header and footer (init.scss and close.scss) @import as expected by the compiler.
Fixed and improved the @import so it works as expected, tested 100%.
Fixed and improved the @mixin so it works as expected, tested 100%.
Added the path to the version script to mk and dev/coverage since it now gets included.
Implemented the conversion of variables appearin in comments.
Added support for MathJAX in our documentation. I put an expression for the computation of the nth-child.
Finished the documentation of the selectors in selector-rules.cpp.
Allow for users to set the 'use empty for undefined variable' instead of generating an error.
Added a NULL_TOKEN so one can define an empty variable.
Added some tests to the compile_declaration() since it could crash the compiler.
Fixed the mark_selector() so nested blocks do get marked as selector blocks.
Avoid the replace_variables() before calling replace_at_keywords() because the @mixin needs to not have its block variables replaced, and the first parameter cannot be a variable.
Added code to allow for component value to be defined in variables. It is still somewhat cluncky though.
Added a TODO.txt file.
Implemented the !global and !default for variables.
Better handling/testing of function arguments at time of declaration.
Correctly support absolute paths in find_file().
Fixed a bug in the lexer for C-like comments: it would never stop if the */ was missing.
Added "null" as an identifier representing false.
Fixed a bug in the parser where finding a variable was considered a terminal token for the current group.
Fixed the installation of the scripts, so now go to the right place (and only the scripts).
Changed the major, minor, and patch version variables to be integers instead of strings.
Fixed the main() function in the tests so we can use all the options without losing track (i.e. 'i' was not properly handled.)
Updated the tests accordingly.