SNAP-187: Started the assembler and tests.
Updated the parser to better handle blocks inside blocks for @media, @document, @supports.
Added a TODO entry about the possibility to list all the possible selector entries in an @selectors so that way one can make sure that only valid tags, classes, identifiers are used.
Broke out the compiler @-keyword expression function so that way we can call it only if necessary.
Implemented a test against @charset and generate an error if not set to UTF-8.
Fixed the replace_if/replace_else so it does not err as much.
Check whether two #id use the same name in the same list of selectors, generate an error if so.
Skip the first space of each C++ comments so we can better align things.
Updated the tests accordingly. Right now they even pump all your memory... working on it. Doing a checkin to make sure I don't lose data.