SNAP-187: Enhanced the assembler some for argument and list handling.
As part of the enhancements, I made changes to the tree of nodes and as a result had to make many changes to the compiler to handle the extra LIST and COMPONENT_VALUE nodes.
Applied a fix to the replace_if() so when followed by an @else, the @else is properly marked as valid at that location even when the @if generates an error.
Added a test in the nodes so we can limit the number of nodes and avoid out of memory errors.
Fixed the handling of the --help command line option, the ++i was necessary there too.
Enhanced the dev/coverage script to be able to run each test one by one with one run of the script.