SNAP-187: Got the assembler at ~99%, the FONT_METRICS is still missing (not implemented in compiler.)
Fixed various problems in the assembler with missing or extra spaces. We still have one with the @-keywords though.
Added coverage logs on a per --test basis.
Moved a few functions around (argify to parser, compile_expression to expression).
Changed the handling of the !important and other "exclamation" flags, the name is inside the EXCLAMATION node and we verify that they appear at the end of declarations.
The exclamation flags are also in nodes now.
Also the parser sees exclamation flags as part of the component value they appear in instead of just declaration which works better with variables.
Fixed the FUNCTION handling by the compiler, the arguments are now argify and not viewed as a simple list of expressions (because some functions expects multiple space separated items as arguments and not just one value or string.)
Deleted various parts that do not apply anymore since I fixed the parser/compiler.
The compiler may find some empty lists while replacing variables and get rid of them.
Fixed bugs with out of range indexes and invalid node types.
Added a few tests to let users know that a #id is not useful but at the start of your selector list, and especially, that two times the same #id is not considered valid.
Fixed a bug in the expression where decimal numbers would be multiplied together using the integer (which was wrong, but the expression class is not yet tested much.)
Added the not() function in the expression class.
Small clean ups here and there (documentation, comments, things moved around, missing B:... for OPEN_CURLYBRACKET, ...)
Updated many of the tests according to all the various changes. They all are 100% except the assembler and the expression test is still missing.