SNAP-187: Implemented the compilation of nested rules and declarations.
Added support for FONT_METRICS.
Included the test to know whether the declaration should watch out for font metrics in the list of values.
Added a test to generate an error if a pseudo-element is not the last selector item (since it is not valid).
Somewhat fixed the definition of the index for the '[]' operator. We still don't have support for a map though.
Properly ignore the calc() and expression() functions in expressions, otherwise it would not come out as expected.
Added a few missing pseudo-classes, functions, and elements (as supported by Firefox).
Fixed the test in the :not() function, if more than one item is present, we have to generate an error.
Fixed the decimal_number_to_string() function which would eventually return an erroneous string (bad precision, missing zeroes).
Ameliorated the test about nested declarations, it properly detects such in more cases as expected.
Added -p option to csspp to specify the decimal numbers precision.
Enhanced errors in declarations to support some Microsoft schemes.
Updated the tests accordingly, we now have 100% coverage of the assembler since we have the FONT_METRICS.