SNAP-187: Added code so the --args option of the csspp command line tool is actually useful.
Added support for arrays and maps in expressions (not well tested yet.)
Fixed a bug with the !exclamation which would inadvertendly skip a token in expressions.
Allow the 'null' token to appear as a unary term in expressions.
Convert identifiers that match a color name into a COLOR node.
Added an identifier() function to convert values in an expression.
Updated the tests accordingly.