SNAP-187: Added many more internal and user functions. Updated corresponding documentation.
Added a main page for the documentation.
Enhanced the get_variable() function so we can force getting the global definition instead of the local one.
Properly added the support for '==' as an equivalent to '=' which was already documented.
Fixed the constant variable names.
Changed all the system function names to use '_' instead of '-' as documented.
Updated the tests accordingly.
Cleaned up the #include.