SNAP-187: Added support for date and time variables (date/time when compiling).
Bumped version to 1.0.2
Fixed link to {less} in compiler reference documentation, also added a link to SASS.
Fixed footer (close.scss) checks using a function so we can make the version all proper.
Fixed a compiler test that end up doing a[b==] and that's equivalent to a[b=] so the error is different than expected in that case.