SNAP-187: Integration of the csspp library in the content plugin.
Took the time to break up the content.cpp file in 9 files. As a result, content.cpp is now "only" 3731 lines.
Removed the content::files::compressor field which we are not using. Instead we have compressor specific field names.
Fixed the SNAP_NAME_CONTENT_FILES_NEW, we had to use an int8_t instead of just 'true'.
Finally decided to move the "javascripts" name in the content get_name() instead of the javascripts plugin.
Added the cpath definitions in the load_attachment() function.
Fixed the process_attachment() to use the file_row directly instead of wasting time with a file_key and reload that key in each function implementation to reach the same row.
Added a fixed-box.js script to handle the case of a box that should not be scrolled within a certain range.