SNAP-187: Added a test to check multiplicative expressions. About 50% done.
Fixed the parenthesis sub-expression computation, I was using the next token.
Better compress output by removing leading zeroes from decimal numbers when not required.
Generate an error if the assembler detects an incompatible dimension (with spaces, *, or /).
Properly apply the decimal number precision when outputting a DECIMAL_NUMBER node in the assembler.
Fixed the assembler output of FONT_METRICS when using PERCENT values.
Added tests to the assembler to check a few things that the previous version would miss.
Test and err when we are about to divide a number by zero.
Implemented support for dimensions in sqrt().
Fixed documentation, some sub-sections names were not taken properly.
Enhanced the expression documentation (better math expressions, explanation, completeness.)