SNAP-187: Added final touches to the multiplicative test, we have 100% coverage.
Added support for several operators on colors (=, !=, +, -, *, /, %).
Added a page documenting the expressions by type of data instead of by operators.
Made a COMPONENT_VALUE compile as expected when no LIST is defined in an OPEN_CURLYBRACKET.
Removed the special handling to support '==' from the expression class. It is in the lexer already. Moved the warning in the lexer.
Final touches on dimension handling in multiplicative operations.
Applied a fix to the power handling to properly accept the value as we wanted to have (and not POWER/POWER!).
Updated the tests accordingly.