SNAP-187: Added tests for relational operators.
Added a function named is_comparable() which we call before is_equal() and is_less_than(). That way we can test compatibility between lhs and rhs in one place.
Fixed the LESS_EQUAL test, it had to use || and not &&.
Added missing relational tests (integer vs decimal number.)
Updated the dev/coverage script and tests accordingly.
Added one test to the assembler test to verify that the '-' works as expected.
Fixed the "coverage" test so it does what it is supposed to do and renamed it "full-coverage".
Small update to the docs.