SNAP-187: Added tests for all the other expression parts.
Added conditional, equality, list, logical and/or, internal function tests.
Fixed the assembler so identifiers that require backslashes are properly output.
Cleaned up the expression by commenting out 3 functions that are not used anymore and made another 5 private.
Changed the expression::boolean() function to return true or false since it already creates an error, we don't need more than that (part of fixing && and ||).
Fixed the equality operators so it calls is_comparable() first.
Fixed a but in the decimal number to string code, it could return "-0" as an answer, instead of just "0".
Fixed the node creation for MAP's. Also changed the management of the comma. Not tested yet.
Properly implement the sin(), cos(), tan() so the parameter is converted to radians before use.
Fixed the acos()/asin()/atan() so the value returned is marked as radians (dimension is "rad").
Changed the red, green, and blue so it returns an integer instead of a decimal number.
Fixed the blue error so it says blue instead of red.
Fixed the cos()/sin()/tan() functions so they do not return a dimension.
Fixed the decimal_number() function so a PERCENT number gets converted to a DECIMAL_NUMBER node as expected.
Removed a few cases in identifier(), string(), integer(), and decimal_number() functions because it cannot get in a function.
Fixed a few set/get decimal_number/integer that were swapped.
Fixed the internal function names, they have to use '_' and not '-'.
Fixed the node to_boolean() function so it properly returns FALSE with the NULL_TOKEN.
Fixed the PERCENT to_string() function by multiplying the percent by 100.
Fixed the tests accordingly.