SNAP-187: Added tests of a few more functions.
Fixed the expression copy. We may want to look into passing the parent expression as a parameter so that way we can copy what is required in the constructor (one place.)
Fixed the function names in error messages (i.e. show name with '_' and not '-'.)
Fixed the text about whether a function exists since we have a sub-LIST now.
Fixed the not() function implementation.
Fixed the number of parameters supported by round().
Fixed the function_exists() function so it returns true for internal functions.