SNAP-187: Got the tests of the release build to run 100%.
Added a call to the csspp_tests in the script building the release so that way we can see that it works.
Applied a real fix to the test for overflow of the integral part of a number.
Applied a few fixes to the unary tests which were making use of if() when it was still bogus.
Fixed two bugs in the lexer test where I would initialize a buffer after I used it.