SNAP-187: Write a test against and fixed the assembler escape_id() function which has to convert characters back from UTF-8 to UTF-32.
Finally changed the set_hsl() and get_hsl() to use radians instead of degrees. That way it directly works with internal functions as it is supposed to.
Fixed the hsl() and hsla() functions so they convert the angle to radiants as expected.
Made sign() accept percentages.
Generate an error on negative numbers when using sqrt().
Added parenthesis so the check for dimensions in sqrt() work as expected.
Removed the adjust_hue(), adjust_saturation(), adjust_lightness() functions, these are external functions at this point.
Fixed the conditional() implementation so the ':' is not taken as a label separator.
Updated the tests accordingly. I'm at 100% except for expressions 94.4%.