SNAP-187: Finished up the coverage tests to include the list (array/map) and reached 100%.
Fixed the post() function index on a map.
Allow identifiers as indexes on a map when accessing it with block-[].
Removed the f_skip_whitespace flag since it was always set to true.
Fixed the set_color() function so identifiers such as "abc" do not get converted to a color when not starting with a hash ("#abc" is a color "abc" is not.)
Removed unused (commented out) functions in the expression class.
Updated a couple other tests accordingly.