SNAP-42: Allow for the deletion of a page using the delete action.
Added code to properly handle the case of a DELETED page.
Added a flag to avoid double calling snap_child::die().
Added a content destroy function to get rid of a page 100%.
Added a corresponding action so a programmer / administrator can delete a page.
Added a permission check for various links (/admin in signature, etc.)
Added a function that returns all links on a page.
Added an action to allow the cleaning up of broken links.
Fixed the server access that would not add '_ajax: 1' with empty requests.
Various small fixes (error descriptions, spaces, documentation, etc.)