SNAP-42: Fixed the lock for the do_update(). We needed to have a #name to avoid lock-ups while updating the root page.
Added a test in the list plugin to verify that entries do get deleted when duplicates are added.
Applied two fixes to the list plugin (1) do not mark a page as modified when we add a "list::..." link; (2) avoid tweaking the listref table without a lock.
Commented out the code I added on last update that would check the last updated date of a list. We have to ignore that test for now because it prevents legitimate updates, somehow.
Added support for a breadcrumb (links to parent pages) which shows up in the output plugin.
Added the libcsspp library version to the admin/versions page.
Fixed the lock in the shorturl plugin, the new lock is specific to each website. The old lock was common to all, but there was no such need.
Added missing throw() in on_backend_action() functions which cannot return if the action passed is invalid.
Added a field in dbutils.
Various cleanups (coding standard, const, docs)...