SNAP-48: Fixed the paypal process since the paypal HTTP header may ask us to close our connection and use a new one for the following request.
Force the connection field and set it to "keep-alive" or "close" depending on what the programmer requested.
Force a connection close if the server replies with a "Connection: close" header field.
Fixed a bug in the get_product_plan() function declaration: the HTTP client object was passed by copy. Also made sure that http_client could not be copied anymore.
Also changed the f_keep_alive default value from false to true.
Also ensures that header values are trimmed of spaces from the end too.