SNAP-187: Implemented the @return in external functions.
Wrote tests for all the external functions. Fixed bugs found on those functions (remove/set unit and adjust_hue() had real problems).
Moved the paths and empty-on-undefined-variable flags to the compiler status to make it easy to implement user defined functions.
Similarly moved a few functions such as find_file() to the compile_state_t class.
Changed the implementation of the conditional() function to allow for invalid operations in the true or false result. Those get ignored and the other result is returned.
Optimized the compile_args() function. It may also be clearer now.
Added the COLOR type as a unary value.
Made the percentage() an internal function. It is otherwise not possible to achieve the capability.
Added external function deg2rad() because hue() returns degrees.
Moved function unique_id() from external to internal functions so it could be implemented.
Added a -c option to the mk script so I can run tests directly that way.
Fixed a problem where argify() ends up being called multiple times. Only the first time are the children argified now.
Removed notice about sourceforge in home page.
Added "opacity" as one of the function that may appear on a filter: ...
Fixed a bug with the minus sign. If not followed by a valid value, it would crash (segmentation fault).
Updated the tests and documentation accordingly.