SNAP-186: Added the timeout tag to the editor to distinguish timeout and auto-reset.
Implemented the timeout in the editor.js file. This represents the duration of the session, opposed to the auto-reset which represents a time when the form should be reset automatically for security reasons.
Note that we will use a minimum of 3 minutes for the timeout and also we don't use the maximum, but instead the maximum minus three. That helps with the possible time difference between the client and server.
This also means the session time to live is set using the timeout value instead of always using 24h.
If the timeout value is defined, then we have to find a corresponding timeout widget otherwise we get an exception.
Fixed a couple of timer identifier that were using -1 instead of NaN (although it should theorically work, it is not specified that -1 would not be a possible return value.)
Fixed the wasModified() function which would not set the modified_ field back to false in case of a restore.
Updated/fixed some documentation.