SNAP-202: Added the cookie-consent-silktide plugin to support cookie consent for Europeans.
Added a couple of fields to dbutils.
Finally added support for int64 type in editor fields.
Added a number (integer/floats) filter with min/max values to the editor. (fixed many comments in there--copy/paste can be a bit off in that situation)
Added a couple of missing PNGs to the CMakeLists.txt file of the output plugin.