SNAP-42: Fixed a bug in the path_info_t which has to clear the f_segments when the path is changed.
Now force a generate_new_lists() and generate_all_lists() call if none happened from the list of entries in the list table. Not absolutely sure whether we need that or not, but it looks like a good idea to me right now.
Fixed a couple of messages in JavaScript that were referencing another function than self.