SNAP-42: Fixed the permission code so when in an IFRAME we avoid saving the IFRAME URL in the redirect.
Fixed the login window so it does not appear in an IFRAME.
Fixed the e-Commerce list of products. It had to use "content::page_type" and not "ecommerce::product" for the link.
Fixed the page type of the /invoices page to ecommerce/invoice instead of epayment/product.
Removed the float64 types from the product names.
Various clean ups.