SNAP-42: Fixed the initialization and conversions of the locale_timezone field.
Actually implemented the default value of the locale_timezone field properly.
Properly implemented the value_to_string() in editor::on_generate_page_content() which was still using the old scheme (all internal...)
Fixed the return value of editor::value_to_string_impl(); it had to return true if none of the internal data types were matched.
Fixed the name of the type: it was not getting set in editor::value_to_string_impl().
Fixed the name of the string type in editor::string_to_value_impl() since the default is not "string" but "".
Added "locale_timezone" as a type to be loaded/saved in a form to the editor DTD. Unfortunately, that's not quite right...