SNAP-43: Added RDFa in our breadcrumbs.
Added access to breadcrumbs settings (i.e. "Home" label, whether to show Home and current page.)
Fixed the cookie consent on_save_editor_fields() to regenerate the JSON settings ONLY when these specific settings get changed.
Actually added no-toolbar as a per widget attribute (the DTD was missing the declaration.)
Switched the info settings to use the editor instead of the form plugin.
Removed a spurious #pragma left behind in the info.cpp.
Added an add_css() for the editor. That fixes a potential problem with dropdowns.
Made the verify_name() function non-static so we can show the key and branch in possible errors.
Moved the privacy-policy page as a child of terms-and-conditions to test breadcrumbs, and it makes (more) sense to have it there.
Fixed the type_theme, it had to be added to the global (content) table and not the branch table.