SNAP-42: Fixed the editor as a few features were not working anymore.
Editor link dialog could not be closed because this was used inside a sub-function.
The link dialog could not be opened because the title is optional and if not defined it would fail the castToString().
Properly position the link dialog depth (z-index) after darkening behind.
Removed the darken "wait" effect.
Fixed the fadeOut of the link dialog (the old version would try to access a now defunct variable member.
Fixed the callbacks. The static table cannot be assigned the callbacks directly. So we do it dynamically.
Fixed the click on a toolbar button, a 'this' parameter was used directly instead of using jQuery(this).
Fixed the call to cancelToolbarHide() which needed to use that and not a static reference.