SNAP-42: Applied a fix for plain and non-plain texts such as the title tag text. Unfortunately, we still have a potential problem with having tags in a title. (TBD)
Fixed the support of "plain" data in the editor. The editor::value_to_string_impl() had to call snap_dom::escape() to be compatible with all the other functions.
Added a COMMAND_SAVE_PLAIN although at this point I'm not convince it is useful.
Fixed the cmd_save() to allow all types (float64, int64, int64_date, xml, plain text) to all accept complex XPath like destinations.
Also added support for a starting '*' in the destination so we can save the same value multiple times in different places.
Fixed a bug in the editor.js which would avoid sending a value in the POST if an error occurred and the user restored the original value (i.e. enter invalid value, try Save, enter original value, try Save, draft version kicks in and value is still considered invalid!)
Fixed the $title variable in the html-header.xsl since we were supposed to use ".../title/node()" and not just ".../title".
Fixed the page number in the pager navigation list of pages. It was starting at 0 instead of 1.
Cleaned up the forgot-password email (display on one line per paragraph, fixed one closing tag.)
Made the emails editable by adiministrators who have the .../website administration permission.
Applied a fix to CMakeLists.txt to allow compiling against ImageMagick 6.
Applied a fix to get log4cplus to use LIBRARIES plurial as expected from a Find....cmake file.
Extended the accepted characters in a URI to include '@', ',', ';', and ':'.
Added support for absolute-min/max length of text in the editor, opposed to the "normal" min/max which ignores tags and entities.
Applied a fix to the editor Save capability because in some situation it would not properly send data to the server.
Fixed a couple of values used by the list paging capability.
Fixed the response of on_error() in the path_error_callback class when the request was an AJAX request. (needed because we cannot call on_redirect() with an error.)
Fixed a call to exit(1) without the "f_snap->" introducer. We could have lost some data by not properly flushing the network connection.
Applied a fix to the server-access.js script in the event the responseXML was not defined, in that case willRedirect() would fail.
Added code to verify that the URI is valid in ajax_redirect().