SNAP-199: Permissions need to be cached otherwise the system is dead slow, so to speak.
Added a string_to_microseconds() to the dbutils so we can use it for such data and cached permission data.
Replaced all the UInt64 to Int64 for time in microseconds since it is supposed to be signed.
Added a security test in the page_redirect() function, which now checks that the HTTP code to redirect is a valid redirect code (301, 302, 303, 307, 308).
Added a cache table to the content plugin.
Various fixes and clean ups.