SNAP-199: Added support for caching CSS and JS files forever (1 year).
Changed the paths so the CSS and JS files are added to the HTML with their version. That way we can use 1 year in the cache duration.
This adds a cache control class so we can setup the Cache-Control field from pages and plugins.
Added a check of the Cache-Control field sent by the client, to take the no-cache and no-store settings in account if present.
Applied a fix (hopefully) to the die() function which would not save the session data.
Switch to using the new cache control functions to generate the output headers.
Fixed the position of the privacy-policy location by moving it below terms-and-conditions in the set_cookie().
The output_result() function now checks whether the Content-Encoding is already set, if so, it does not try to compress the data. This allows us to return the pre-compressed CSS and JS files.
Changed the snap version validation by using a sub-validation function (validate_basic_name).
Applied a fix to the validate_operator() so the 'op' parameter is always reset even if operator string is invalid.
Removed the test of namespace for the browser specification because we can now directly call the validate_basic_name().
Fixed the test of md5 buffer size, instead of checking whether it is null, we verify that the size is 16 bytes.
Added the nosniff content type option for Internet Explorer and Chrome to avoid cross domain scripts and CSS.
For the CSS and JS files, we now check minified, compressed, and uncompressed versions.
Fixed the force_extended_revision() so it correctly reset the branch and revision.
Marked a few pages with "no-store" (e-payment and users).