SNAP-42: Changed the list handling to include a priority.
Updated the list loop handler which did not need to go through all the sites defined in there since we cannot handle but site_key.
Added a way for the list system to automatically re-add all the pages of a website to re-parse them all over and over again. This uses a really high (slow) priority.
Applied a few optimization to the list implementation.
Fixed the catch(...) so it logs the fact that an error occured.
Enhanced the f_updating flag handling with an RAII class to make sure it gets cleared.
Gave access to the f_updating flag so the list can switch priority when updating opposed to other updates.
Made a change to the content::modified_content_impl() so the modification data is also saved in the content table.
Changed the key name in the dbutils::get_column_name() function for the list table so we show the priority.
Fixed the log4cplus configuration file to keep 10 backups instead of just 1 by default.