SNAP-49: Changed the latency to include it when adding items to the list.
Added support for the programmer to change the amount of time to wait before handling certain pages (i.e. the list processing latency). The latency is now part of the on_modified_content() instead of the loop, in effect avoiding the 10 seconds latency problem.
Changed the loop to make use of the smallest timestamp for the next wait.
Fixed the start_date handling in list::on_modified_content() -- there is the process start_date and the key_start_date. This is very important to properly drop the row we want to drop.
When using a priority and a date and re-modifying a row before it was worked on by the list plugin, we must keep the smallest priority and largest timestamp.
Changed the generate_new_lists() to read the list of new lists and then handle them, to make sure the links_ctxt does not re-reset the links table over and over again between each call.
Added a (probably useless) backend action to reset a site by re-add all the pages to the list table. This makes use of a very slow priority.
Added a type in dbutils to support the new format of rows in listref.
Removed the extra #ref in the listref table, at first we were using the list table but decided that parameters could be much better optimized per table.
Enhanced / fixed some documentation.
Applied a "fix" (added a delay of 1us) so the last modification time of permissions::on_modified_link() to avoid some cache problems between lists and main processes.
Fixed the "resetlists" action name, the final 's' was missing.