SNAP-140: Further fixed problems with the caches, not only the permissions (cache table) but also the sites table.
Actually moved the clearing of the libQtCassandra caches outside of the if() even if it would work for us, it is important when installing a new system.
Fixed the priority / timestamp test, it was supposed to continue when the priority is less than LIST_PRIORITY_SLOW.
Changed the timestamp limit of the permission cache to use EXPECTED_TIME_ACCURACY_EPSILON which at this time is 10ms.