SNAP-42: Looking at a hit from Baidu, our shorturl was generating errors.
I changed the redirect behavior to make use of the new signal check_for_redirect().
Use 301 for the redirect (instead of 302). Although 302 is good to allow for changing the redirect to another location, it breaks SEO juice.
When a redirect fails because the page does not exist, we instead show page "s" and include an error message.
Various clean ups.