SNAP-187: Integrated the csspp changes (1.0.7) to Snap core. This makes the editor toolbar work as expected.
Fixed the content::load_attachment() so it can load a CSS or JS with version and .min. in the filename.
Fixed dbutils::string_to_key() so it can parse an MD5 to a buffer without the need for spaces between each hex.
Added a new action (newfile) to repeat a file check.