EX-1: Fixed the server_access and taxonomy to remove loops.
Changed the taxonomy content.xml to force the owner to "output" so taxonomy does not have to depend on the layout or path plugins.
Added missing directory to doxy file so we see all plugins in our documentation graphs.
Many #include cleaned up to streamline the library and some plugins.