EX-30: Added code to save the last user session in the user account.
If logging in again and their is a last user session, by default, we "delete" the session (we force a log out, really).
Added support for administrators to allow more than one login.
Fixed a potential security issue when a user verifies his account and another user was logged in. We should look into whether we could call the logout function instead of having a semi-duplicate.
Added a page for users settings, only I did not program any of it. Also I may want to either do it in the info plugin or have a users_settings plugin.