EX-1: Started integrating snap_communicator.
Replaced the listen() function with the new code using the snap_communicator class.
Fixed a potential security issue in snap_child::process() where the socket could be missed and never closed.
Added a name to the snap connection implementation so we can show it in errors.
Fixed a problem with checking the socket validity. For timers, the socket remains -1.
Fixed the set_timeout() function, it cannot call the attach_event() if the event was not yet created.
Fixed several virtual functions in snap_communicator classes: get_events() was missing the "get_" part.
Fixed a potential problem with creating a snap_server_client_connection with a negative socket number.
Now we close the snap_server_client_connection socket if not -1.
Added support for the SO_KEEPALIVE flag in the snap_communicator environment.
Made the snap_communicator run() function virtual.
Noticed a couple of cases where the f_socket in the tcp_server constructor could end up lost.
Added a user permission so /user is always accessible to website administrators.
Fixed the "delete session if double sessions are found" in case the session is your own in the same browser (i.e. your session timed out and you have to re-login, then the session does not change, but you still have a previous login session field...)
Fixed the error / warning CSS, the warning was supposed to appear before the error so in case of an error we see the error color.
Various clean ups.