EX-17: Added code to allow a sub-image (i.e. resized, with shadow, reflection, etc.) to make use of the main image when not yet available.
Fixed a test of an MD5 key. It has to have a size of exactly 16 bytes instead of not empty.
It looks like the path::add_restore_link_to_signature_for() function was using restore_error_callback.has_error() without the '!' operator and that was most certainly wrong.
Various clean ups.