SNAP-150: Added code to support proper dates in Atom and RSS.
Fixed the code so we get the correct URL in all entries shown in a RSS.
Fixed the list generation to (1) have a unique key and (2) use the finball::created of the main page instead of the revision.
Added code to support YYYY-MM-DD as a date format in string_to_date().
Fixed a really bad bug in the call_format() where the variable index was not being incremented.
Added a preg_replace() function to our expression support.