EX-15: "Browse" button added to the editor widgets that accept at least one attachment.
Added a "browse" attribute to the "attachment" tag of the editor widgets so the programmer may prevent the Browse button from appearing.
Added CSS for our Browse button (Core only version at this time).
Added the code to give an error when the file format cannot be determined.
Various cleanups.