SNAP-150: Changed the implementation of the filter_text() signal to implement the filter conversion in the signal instead of the token_filter() function.
Made the implementation of the insert_html_string_to_xml_doc() similar to the replace.
Changed the replace_token() signature, now the owner is saved in the ipath as "token_owner".
Fixed a potential security issue in hashtag that would make use of the title as is in an attribute.
Fixed the feed.cpp as the main URI had to be restored to "/".
Various cleanups.