EX-14: Actually call the trash_page() function when adding an attachment and another attachment already exists (and we are overwriting the previous one.)
Changed the create_content() and create_attachment() so if the path references a deleted page, it gets undeleted (however, it does not get unhidden if currently hidden.)
Added that specific case so we can now got from DELETED back to NORMAL (if NOT_WORKING).
Fixed the display (dbutils) of content::files::created/modified dates which are time_t values and not int64_t microsecnds.
Fixed a security issue with image previews, these could be viewed even if the page itself had been deleted.
Added "Sans-Serif" for our errors so that way the font looks good (at least to my point of view it looks better this way).
Added a "delete all attachments" function to the attachment plugin, although at this time I am not using it (but it was tested and worked just fine.)
Added the new tables (cache, test_results, tracker) to the snapdb tool so they get deleted as expected.
Started a function to generate a teaser.
Added some documentation.
Various clean ups.